% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{col_df}
\alias{col_df}
\title{scolor scheme of sequence aligned}
\format{A data frame with 24 rows and 6 variables}
\usage{
data(col_df)
}
\description{
A dataset containing the color scheme of 24 observations of 6
variables. The variables are as follows:
}
\details{
\itemize{
  \item{\code{category}}{a factor with levels \code{Aromatic} \code{Cysteines} \code{Glycines} \code{Hydrophobic} \code{Negative charge} \code{Polar} \code{Positive charge} \code{Prolines} \code{Unconserved}}
  \item{\code{colour}}{a character vector}
  \item{\code{re_position}}{a character vector}
  \item{\code{thred}}{a numeric vector}
  \item{\code{re_gp}}{a character vector}
  \item{\code{type}}{a character vector}
}
}
\keyword{datasets}
