\name{sampleseq}
\alias{sampleseq}
\docType{data}
\title{the data frame of sample sequence aligned }
\description{
This data set gives 9 obs. of 50 variables containing 9 sequences aligned.
}
\usage{data("sampleseq")}
\format{
  A data frame with 9 observations on the following 50 variables.
  \describe{
    \item{\code{V164}}{a factor with levels \code{-} \code{i} \code{q}}
    \item{\code{V165}}{a factor with levels \code{-} \code{a} \code{t} \code{v}}
    \item{\code{V166}}{a factor with levels \code{-} \code{c} \code{l} \code{y}}
    \item{\code{V167}}{a factor with levels \code{-} \code{n} \code{r}}
    \item{\code{V168}}{a factor with levels \code{-} \code{a} \code{p} \code{y}}
    \item{\code{V169}}{a factor with levels \code{-} \code{d} \code{r}}
    \item{\code{V170}}{a factor with levels \code{-} \code{f} \code{h} \code{w}}
    \item{\code{V171}}{a factor with levels \code{-} \code{g} \code{t}}
    \item{\code{V172}}{a factor with levels \code{-} \code{i} \code{l} \code{m} \code{q} \code{v}}
    \item{\code{V173}}{a factor with levels \code{a} \code{d} \code{e} \code{p}}
    \item{\code{V174}}{a factor with levels \code{i} \code{n} \code{q}}
    \item{\code{V175}}{a factor with levels \code{d} \code{g} \code{p}}
    \item{\code{V176}}{a factor with levels \code{f} \code{l} \code{r} \code{v} \code{w} \code{y}}
    \item{\code{V177}}{a factor with levels \code{d} \code{e} \code{h} \code{i} \code{v}}
    \item{\code{V178}}{a factor with levels \code{-} \code{e} \code{h} \code{r} \code{t}}
    \item{\code{V179}}{a factor with levels \code{y}}
    \item{\code{V180}}{a factor with levels \code{m} \code{p} \code{s} \code{t}}
    \item{\code{V181}}{a factor with levels \code{a} \code{d} \code{e} \code{q}}
    \item{\code{V182}}{a factor with levels \code{a} \code{e} \code{t}}
    \item{\code{V183}}{a factor with levels \code{d} \code{e}}
    \item{\code{V184}}{a factor with levels \code{h} \code{k} \code{q} \code{r}}
    \item{\code{V185}}{a factor with levels \code{a} \code{d} \code{k} \code{q}}
    \item{\code{V186}}{a factor with levels \code{t} \code{v}}
    \item{\code{V187}}{a factor with levels \code{w}}
    \item{\code{V188}}{a factor with levels \code{a} \code{g} \code{i} \code{n} \code{r}}
    \item{\code{V189}}{a factor with levels \code{t}}
    \item{\code{V190}}{a factor with levels \code{l} \code{v}}
    \item{\code{V191}}{a factor with levels \code{c} \code{f} \code{i} \code{y}}
    \item{\code{V192}}{a factor with levels \code{d} \code{e} \code{k} \code{q} \code{r} \code{t}}
    \item{\code{V193}}{a factor with levels \code{r} \code{t}}
    \item{\code{V194}}{a factor with levels \code{l} \code{q}}
    \item{\code{V195}}{a factor with levels \code{c} \code{e} \code{k} \code{l} \code{t}}
    \item{\code{V196}}{a factor with levels \code{a} \code{d} \code{k} \code{s}}
    \item{\code{V197}}{a factor with levels \code{l} \code{m} \code{v}}
    \item{\code{V198}}{a factor with levels \code{i} \code{l} \code{t} \code{y}}
    \item{\code{V199}}{a factor with levels \code{e} \code{h} \code{k} \code{p} \code{r}}
    \item{\code{V200}}{a factor with levels \code{g} \code{k} \code{t}}
    \item{\code{V201}}{a factor with levels \code{h} \code{l} \code{r}}
    \item{\code{V202}}{a factor with levels \code{a}}
    \item{\code{V203}}{a factor with levels \code{c} \code{h}}
    \item{\code{V204}}{a factor with levels \code{d} \code{h} \code{q} \code{y}}
    \item{\code{V205}}{a factor with levels \code{e} \code{s}}
    \item{\code{V206}}{a factor with levels \code{f} \code{h} \code{y}}
    \item{\code{V207}}{a factor with levels \code{l} \code{m} \code{n}}
    \item{\code{V208}}{a factor with levels \code{d} \code{e} \code{h} \code{q} \code{r}}
    \item{\code{V209}}{a factor with levels \code{g} \code{i}}
    \item{\code{V210}}{a factor with levels \code{f} \code{i} \code{l} \code{v}}
    \item{\code{V211}}{a factor with levels \code{-} \code{p}}
    \item{\code{V212}}{a factor with levels \code{-} \code{l}}
    \item{\code{V213}}{a factor with levels \code{-} \code{l}}
  }
}
\details{a sample referenece}
\source{a sample referenece}
\references{a sample referenece}
\examples{
data(sampleseq)
## maybe str(sampleseq) ; plot(sampleseq) ...
}
\keyword{datasets}
