\name{testseq}
\alias{testseq}
\docType{data}
\title{
the data frame of test sequence aligned
}
\description{
This data set gives 9 obs. of 7 variables containing 9 sequences aligned.
}
\usage{data("testseq")}
\format{
  A data frame with 9 observations on the following 7 variables.
  \describe{
    \item{\code{V180}}{a factor with levels \code{m} \code{p} \code{s} \code{t}}
    \item{\code{V181}}{a factor with levels \code{a} \code{d} \code{e} \code{q}}
    \item{\code{V182}}{a factor with levels \code{a} \code{e} \code{t}}
    \item{\code{V183}}{a factor with levels \code{d} \code{e}}
    \item{\code{V184}}{a factor with levels \code{h} \code{k} \code{q} \code{r}}
    \item{\code{V185}}{a factor with levels \code{a} \code{d} \code{k} \code{q}}
    \item{\code{V186}}{a factor with levels \code{t} \code{v}}
  }
}
\details{a test reference}
\source{a test reference}
\references{a test reference}
\examples{
data(testseq)
## maybe str(testseq) ; plot(testseq) ...
}
\keyword{datasets}
