\name{testseqcolor}
\alias{testseqcolor}
\docType{data}
\title{
the data frame of color related to test sequence aligned
}
\description{
This dataset give the color related to test sequence aligned.
}
\usage{data("testseqcolor")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
mapping test sequence aligned with color
}
\source{
a reference which is the outcome of mapping test sequence aligned to a color matrix
}
\references{
a reference
}
\examples{
data(testseqcolor)
## maybe str(testseqcolor) ; plot(testseqcolor) ...
}
\keyword{datasets}
