\name{sampleseqcolor}
\alias{sampleseqcolor}
\docType{data}
\title{
the data frame of color related to sample sequence aligned
}
\description{
This dataset give the color related to sample sequence aligned.
}
\usage{data("sampleseqcolor")}
\format{
  A data frame with 0 observations on the following 2 variables.
  \describe{
    \item{\code{x}}{a numeric vector}
    \item{\code{y}}{a numeric vector}
  }
}
\details{
mapping sample sequence aligned with color
}
\source{
a reference which is the outcome of mapping sample sequence aligned to a color matrix
}
\references{
a reference
}
\examples{
data(sampleseqcolor)
## maybe str(sampleseqcolor) ; plot(sampleseqcolor) ...
}
\keyword{datasets}
