% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{testseq}
\alias{testseq}
\title{test sequence aligned}
\format{A data frame with 9 rows and 7 variables}
\usage{
data(testseq)
}
\description{
A dataset containing the test sequence of 9 observations of 7
variables. The variables are as follows:
}
\details{
\itemize{
   \item{\code{V180}}{a factor with levels \code{m} \code{p} \code{s} \code{t}}
   \item{\code{V181}}{a factor with levels \code{a} \code{d} \code{e} \code{q}}
   \item{\code{V182}}{a factor with levels \code{a} \code{e} \code{t}}
   \item{\code{V183}}{a factor with levels \code{d} \code{e}}
   \item{\code{V184}}{a factor with levels \code{h} \code{k} \code{q} \code{r}}
   \item{\code{V185}}{a factor with levels \code{a} \code{d} \code{k} \code{q}}
   \item{\code{V186}}{a factor with levels \code{t} \code{v}}
}
}
\keyword{datasets}
