% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/documention.r
\docType{data}
\name{testseqcolor}
\alias{testseqcolor}
\title{color related to test sequence aligned}
\format{A data frame with 0 observations on the following 2 variables.}
\usage{
data(testseqcolor)
}
\description{
This dataset give the color related to test sequence aligned.
}
\details{
\itemize{
   \item{\code{x}}{a numeric vector}
   \item{\code{y}}{a numeric vector}
}
}
\keyword{datasets}
